split out rack middleware from exception notifier core
