+++ lib/exception_notification.rb
@@ -1,1 +1,2 @@
 require 'exception_notifier'
+require 'exception_notification/rack'
